allow users to select clickhouse week mode
Currently, the cube.js ClickHouse adapter hard codes the start of week
to be the Monday of the selected dimension's week. Instead, use the
"toStartOfWeek" clickhouse function which takes a "mode" parameter that
indicates whether the week should start on Sunday or Monday. Expose
this parameter as an environment variable to cube.js so that users can
choose whether the week should begin on a Sunday or a Monday.